Move Embedded to ReifyQuotes
